+++ packages/cli/resources/functions/fileExists.js
@@ -1,7 +1,0 @@
-const fs = require('fs');
-
-module.exports = (targetValue) => {
-    if (!fs.existsSync(targetValue)) {
-        return [{ message: `file "${targetValue}" does not exist` }];
-    }
-}