+++ packages/cli/src/initialise/initialise.ts
@@ -1,6 +1,6 @@
+import { checkoutIntegrations } from './checkout-integrations';
 import { Context } from '../types';
 import { initialiseBuildDirectory } from './initialise-build-directory';
-import { checkoutIntegrations } from './checkout-integrations';
 
 export const initialise = async (context: Context): Promise<void> => {
     await checkoutIntegrations(context);