+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### 0.3.5 - 2021-10-23
+### Changed
+ - CLI: Extracted config linting logic to a separate package.
+ - Updated dependencies.
+
 ## 0.3.4 - 2021-08-20
 ### Fixed
  - CLI: Fixed bad regex being used for detecting relative path in simple volume definition.